begin rewrite: add lionkor/commandline
