+++ .gitmodules
@@ -1,0 +1,6 @@
+[submodule "commandline"]
+	path = commandline
+	url = https://github.com/lionkor/commandline
+[submodule "include/commandline"]
+	path = include/commandline
+	url = https://github.com/lionkor/commandline
