--- include/Compressor.h
@@ -1,11 +1,0 @@
-// Copyright (c) 2019-present Anonymous275.
-// BeamMP Server code is not in the public domain and is not free software.
-// One must be granted explicit permission by the copyright holder in order to modify or distribute any part of the source or binaries.
-// Anything else is prohibited. Modified works may not be published and have be upstreamed to the official repository.
-///
-/// Created by Anonymous275 on 7/24/2020
-///
-#pragma once
-#include <string>
-std::string Comp(std::string Data);
-std::string DeComp(std::string Compressed);
