+++ include/TConsole.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <atomic>
+#include <commandline/commandline.h>
+
+class TConsole {
+public:
+    TConsole();
+
+private:
+    Commandline _Commandline;
+};
