--- src/Network/NetMain.cpp
@@ -1,9 +1,0 @@
-#include "Network.h"
-#include <memory>
-#include <thread>
-std::unique_ptr<ClientInterface> CI;
-void NetMain() {
-    std::thread TCP(TCPServerMain);
-    TCP.detach();
-    UDPServerMain();
-}
